+++ dtaianomaly/anomaly_detection/__init__.py
@@ -20,6 +20,7 @@ from .BaseNeuralDetector_utils import (
     TimeSeriesDataset,
 )
 from .ClusterBasedLocalOutlierFactor import ClusterBasedLocalOutlierFactor
+from .ConvolutionalNeuralNetwork import ConvolutionalNeuralNetwork
 from .CopulaBasedOutlierDetector import CopulaBasedOutlierDetector
 from .DWT_MLEAD import DWT_MLEAD
 from .HistogramBasedOutlierScore import HistogramBasedOutlierScore
@@ -29,6 +30,7 @@ from .KMeansAnomalyDetector import KMeansAnomalyDetector
 from .KNearestNeighbors import KNearestNeighbors
 from .KShapeAnomalyDetector import KShapeAnomalyDetector
 from .LocalOutlierFactor import LocalOutlierFactor
+from .LongShortTermMemoryNetwork import LongShortTermMemoryNetwork
 from .MatrixProfileDetector import MatrixProfileDetector
 from .MedianMethod import MedianMethod
 from .MultilayerPerceptron import MultilayerPerceptron
@@ -37,6 +39,7 @@ from .OneClassSupportVectorMachine import OneClassSupportVectorMachine
 from .PrincipalComponentAnalysis import PrincipalComponentAnalysis
 from .PyODAnomalyDetector import PyODAnomalyDetector
 from .RobustPrincipalComponentAnalysis import RobustPrincipalComponentAnalysis
+from .Transformer import Transformer
 from .windowing_utils import (
     check_is_valid_window_size,
     compute_window_size,
@@ -84,4 +87,7 @@ __all__ = [
     "TimeSeriesDataset",
     "BaseNeuralForecastingDetector",
     "BaseNeuralReconstructionDetector",
+    "LongShortTermMemoryNetwork",
+    "ConvolutionalNeuralNetwork",
+    "Transformer",
 ]