+++ tests/anomaly_detection/test_detectors.py
@@ -53,6 +53,7 @@ def initialize(cls):
         'neighborhood_size_before': 15,
         'detector': anomaly_detection.IsolationForest(window_size=15),
         'preprocessor': preprocessing.Identity(),
+        'n_epochs': 1
     }
     sig = inspect.signature(cls.__init__)
     accepted_params = set(sig.parameters) - {"self"}